+++ docs/dev-logs/2026-06-06-frontend-common-ui-split-bundle-v4.md
@@ -1,0 +1,13 @@
+﻿# Frontend common UI split bundle v4
+
+## Summary
+- `PublicHomePageSections`를 hero/course 파일로 분리
+- `QuizGenPageSections`를 hero/picker/condition 파일로 분리
+- `AdminUsersPageSections`를 header/filter/table/mobile 카드로 분리
+
+## Validation
+- pending
+
+## Notes
+- 페이지 import 경로는 유지했다
+- 역할 필터/강의 선택/퀴즈 조건 UI는 그대로 보존했다
